requirements.txt
snowflake-snowpark-python>=1.18.0
streamlit>=1.35.0
pypdf>=4.0.0
python-dotenv>=1.0.0
